+++ akshare/__init__.py
@@ -3231,10 +3231,23 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.18.23 fix: fix fund_fee_em interface
 1.18.24 fix: fix stock_individual_spot_xq interface
 1.18.25 fix: fix qdii_e_comm_jsl interface
+1.18.26 fix: fix fund_fh_em interface
 """
 
 from akshare._version import __version__
 __author__ = "AKFamily"
+
+"""
+
+"""
+from akshare.futures.futures_settle import (
+    futures_settle,
+    futures_settle_gfex,
+    futures_settle_ine,
+    futures_settle_czce,
+    futures_settle_cffex,
+    futures_settle_shfe,
+)
 
 """
 国债收益率