+++ akshare/stock_feature/stock_gdfx_em.py
@@ -275,6 +275,7 @@ def stock_gdfx_free_holding_change_em(date: str = "20210930") -> pd.DataFrame:
         "-",
         "-",
         "-",
+        "-",
     ]
     big_df = big_df[
         [